separate max duration to mixin + spec helper
