+++ app/models/session.rb
@@ -1,14 +1,16 @@
 # frozen_string_literal: true
 
 require 'common/models/redis_store'
+require 'common/models/concerns/maximum_redis_lifetime'
 
 class Session < Common::RedisStore
+  include Common::MaximumRedisLifetime
+
   redis_store REDIS_CONFIG['session_store']['namespace']
   redis_ttl REDIS_CONFIG['session_store']['each_ttl']
   redis_key :token
 
   DEFAULT_TOKEN_LENGTH = 40
-  MAX_SESSION_LIFETIME = 12.hours
 
   attribute :token
   attribute :uuid
@@ -18,16 +20,12 @@ class Session < Common::RedisStore
   validates :uuid, presence: true
   validates :created_at, presence: true
 
-  validate :within_maximum_ttl
+  # validate :within_maximum_ttl
 
   after_initialize :setup_defaults
 
   def self.obscure_token(token)
     Digest::SHA256.hexdigest(token)[0..20]
-  end
-
-  def expire(ttl)
-    (ttl < maximum_time_remaining) ? super(ttl) : super(maximum_time_remaining)
   end
 
   private
@@ -44,15 +42,5 @@ class Session < Common::RedisStore
   def setup_defaults
     @token ||= secure_random_token
     @created_at ||= Time.now.utc
-  end
-
-  def maximum_time_remaining
-    (@created_at + MAX_SESSION_LIFETIME - Time.now.utc).round
-  end
-
-  def within_maximum_ttl
-    if maximum_time_remaining.negative?
-      errors.add(:created_at, "is more than the max of [#{MAX_SESSION_LIFETIME}] ago. Session is too old")
-    end
   end
 end
