+++ spec/models/session_spec.rb
@@ -1,10 +1,15 @@
 # frozen_string_literal: true
 
 require 'rails_helper'
+require 'lib/common/maximum_redis_lifetime_spec_helper'
 
 RSpec.describe Session, type: :model do
   let(:attributes) { { uuid: 'abcd-1234' } }
   subject { described_class.new(attributes) }
+
+  it_behaves_like 'a redis store with a maximum lifetime' do
+    subject { described_class.new(attributes) }
+  end
 
   context 'session without attributes' do
     it 'expect ttl to an Integer' do
@@ -32,34 +37,6 @@ RSpec.describe Session, type: :model do
       it 'extends a session' do
         expect(subject.expire(7200)).to eq(true)
         expect(subject.ttl).to eq(7200)
-      end
-      it 'will extend the session only up to the maximum ttl' do
-        Timecop.freeze(Time.current)
-        subject.created_at = subject.created_at - (described_class::MAX_SESSION_LIFETIME - 10.seconds)
-        expect(subject.expire(3600)).to eq(true)
-        expect(subject.ttl).to eq(10)
-        Timecop.return
-      end
-      it 'allows for continuous session extension up to the maximum' do
-        start_time = Time.current
-        Timecop.freeze(start_time)
-
-        # keep extending session so Redis doesn't kill it while remaining
-        # within Sesion::MAX_SESSION_LIFETIME
-        increment = subject.redis_namespace_ttl - 1.minute
-        max_hours = described_class::MAX_SESSION_LIFETIME / 1.hour
-        (1...max_hours).each do |hour|
-          Timecop.freeze(start_time + increment * hour)
-          expect(subject.expire(described_class.redis_namespace_ttl)).to eq(true)
-          expect(subject.ttl).to eq(described_class.redis_namespace_ttl)
-        end
-
-        # still within Session::MAX_SESSION_LIFETIME by 720 seconds
-        Timecop.freeze(start_time + increment * max_hours)
-        expect(subject.expire(described_class.redis_namespace_ttl)).to eq(true)
-        expect(subject.ttl).to eq(720)
-
-        Timecop.return
       end
     end
